requirements.txt
appopener
beautifulsoup4
email-validator
feedparser
numpy
pyautogui
pyjokes
PyAudio
python-dotenv
requests
scikit-learn
SpeechRecognition
stability-sdk
tensorflow
twilio
WMI
wolframalpha
youtube-search-python
pydub
pyttsx3
speedtest-cli